# Make R class scanner faster and memory efficient