--- a/butterknife-compiler/src/main/java/butterknife/compiler/Id.java
+++ b/butterknife-compiler/src/main/java/butterknife/compiler/Id.java
@@ -2,29 +2,38 @@ package butterknife.compiler;
 
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
+import com.sun.tools.javac.code.Symbol;
 
 /**
  * Represents an ID of an Android resource.
  */
 final class Id {
   private static final ClassName ANDROID_R = ClassName.get("android", "R");
+  private static final String R = "R";
 
   final int value;
   final CodeBlock code;
   final boolean qualifed;
 
   Id(int value) {
-    this.value = value;
-    this.code = CodeBlock.of("$L", value);
-    this.qualifed = false;
+    this(value, null);
   }
 
-  Id(int value, ClassName className, String resourceName) {
+  Id(int value, Symbol rSymbol) {
     this.value = value;
-    this.code = className.topLevelClassName().equals(ANDROID_R)
-      ? CodeBlock.of("$L.$N", className, resourceName)
-      : CodeBlock.of("$T.$N", className, resourceName);
-    this.qualifed = true;
+    if (rSymbol != null) {
+      ClassName className = ClassName.get(rSymbol.packge().getQualifiedName().toString(), R,
+          rSymbol.enclClass().name.toString());
+      String resourceName = rSymbol.name.toString();
+
+      this.code = className.topLevelClassName().equals(ANDROID_R)
+        ? CodeBlock.of("$L.$N", className, resourceName)
+        : CodeBlock.of("$T.$N", className, resourceName);
+      this.qualifed = true;
+    } else {
+      this.code = CodeBlock.of("$L", value);
+      this.qualifed = false;
+    }
   }
 
   @Override public boolean equals(Object o) {