… butterknife-compiler/src/main/java/butterknife/compiler/ButterKnifeProcessor.java
… butterknife-compiler/src/main/java/butterknife/compiler/QualifiedId.java