The current R class scanner scans all R classes ahead of time, while discarding non referenced fields. This leads to a lot of wasted time and memory pressure during the build. By moving the scanning to happen on demand, both these issues can be mitigated. This also makes the logic much simpler.

Compiler Memory Usage
==================
| Before | After |
| :----: |:----:|
| <img width="580" alt="screen shot 2018-04-09 at 11 49 03 pm" src="https://user-images.githubusercontent.com/291148/38540660-86d04ec4-3c51-11e8-879c-d4d6bfd95fa0.png"> | <img width="531" alt="screen shot 2018-04-09 at 11 23 43 pm" src="https://user-images.githubusercontent.com/291148/38540677-96992ec0-3c51-11e8-8e71-4efb2cf4f484.png"> |
| <img width="1015" alt="screen shot 2018-04-09 at 11 48 55 pm" src="https://user-images.githubusercontent.com/291148/38540673-922fe5a4-3c51-11e8-96bc-c7af1f46fd69.png"> | <img width="390" alt="screen shot 2018-04-09 at 11 23 49 pm" src="https://user-images.githubusercontent.com/291148/38540747-c632c3f8-3c51-11e8-92e6-5af827ea77e6.png"> |

No significant memory retained by Butterknife anymore

Time spent in ButterKnifeCompiler
=========================
| Before | After |
| :----: |:----:|
| <img width="1085" alt="screen shot 2018-04-09 at 11 44 42 pm" src="https://user-images.githubusercontent.com/291148/38540815-05fc48c4-3c52-11e8-928c-dc73cf28e82c.png"> |<img width="1017" alt="screen shot 2018-04-09 at 11 37 29 pm" src="https://user-images.githubusercontent.com/291148/38540825-0d43b130-3c52-11e8-9f36-9926d0aaf2e4.png"> |

Down from **6.5s to 1.3s, a 5x improvement**